% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poems.R
\docType{data}
\name{poems}
\alias{poems}
\title{Best Persian Poems}
\format{
A tibble with 54524 rows and 5 variables:
}
\source{
\href{https://github.com/mahmoud-eskandari/PersianBooksDataset}{Mahmud Eskandari}
}
\usage{
poems
}
\description{
The dataset contains poems from more that 60 famous Persian poets scraped from ganjoor.net
}
\keyword{datasets}
